+++ orchestrator/src/formatting_orchestrator.rs
@@ -1,5 +1,3 @@
-use std::io::{IsTerminal, Read};
-
 use crate::file_formatter::FileFormatter;
 use log::LevelFilter;
 
@@ -15,14 +13,7 @@ impl FormattingOrchestrator {
     pub fn run<T: FormatterConfiguration>(file_formatter: FileFormatter, config: T) {
         let files = config.get_paths();
         if files.is_empty() {
-            let mut input = String::new();
-            if std::io::stdin().is_terminal() {
-                eprintln!("waiting for stdin...");
-            }
-            let mut stdin = std::io::stdin().lock();
-
-            stdin.read_to_string(&mut input).unwrap();
-            file_formatter.format_stdin_to_stdout(&input);
+            file_formatter.format_stdin_to_stdout();
         } else if config.is_verify() {
             file_formatter.check_files(&files);
         } else if config.is_write() {